Xong chuong 3

+++ Chuong3/21110660_level_02_01.cpp
@@ -19,7 +19,7 @@ int main()
     }
     else
     {
-        double ketqua = float(a) / b;
+        double ketqua = double(a) / b;
         cout << ketqua;
     }
